test(ConfigParser): リファクタリングに向けて、テストファイルを集約
