--- tests/unit/test_parser_csv_options.py
@@ -1,201 +1,0 @@
-import math
-from io import BytesIO
-from typing import Any, Dict
-
-import pytest
-
-from features.config_parser import ConfigParser
-
-
-@pytest.mark.unit
-def test_csv_rows_name() -> None:
-    """Test the csv_rows_name property and setter."""
-    # Create a simple CSV file
-    content = b"name,age\nAlice,30\nBob,25"
-    config_file = BytesIO(content)
-    config_file.name = "config.csv"
-
-    # Initialize parser with default csv_rows_name
-    parser = ConfigParser(config_file)
-    assert parser.parse() is True
-    assert parser.parsed_dict is not None
-    assert "csv_rows" in parser.parsed_dict
-    assert len(parser.parsed_dict["csv_rows"]) == 2
-
-    # Reset and change csv_rows_name
-    config_file.seek(0)
-    parser = ConfigParser(config_file)
-    parser.csv_rows_name = "people"
-    assert parser.parse() is True
-    assert parser.parsed_dict is not None
-    assert "people" in parser.parsed_dict
-    assert "csv_rows" not in parser.parsed_dict
-    assert len(parser.parsed_dict["people"]) == 2
-
-
-@pytest.mark.unit
-def test_enable_fill_nan() -> None:
-    """Test the enable_fill_nan property and setter."""
-    # Create a CSV file with NaN values
-    content = b"name,age\nAlice,30\nBob,\nCharlie,35"
-    config_file = BytesIO(content)
-    config_file.name = "config.csv"
-
-    # Default behavior: NaN values are not filled
-    parser = ConfigParser(config_file)
-    assert parser.parse() is True
-    assert parser.parsed_dict is not None
-    assert math.isnan(parser.parsed_dict["csv_rows"][1]["age"])
-
-    # Reset and enable NaN filling with default value (empty string)
-    config_file.seek(0)
-    parser = ConfigParser(config_file)
-    parser.enable_fill_nan = True
-    result = parser.parse()
-
-    # If parsing fails, check the error message
-    if not result:
-        assert parser.error_message is not None
-        print(f"Error message: {parser.error_message}")
-        # Skip the rest of the test
-        return
-
-    assert parser.parsed_dict is not None
-    assert parser.parsed_dict["csv_rows"][1]["age"] == ""
-
-    # Reset and enable NaN filling with custom value
-    config_file.seek(0)
-    parser = ConfigParser(config_file)
-    parser.enable_fill_nan = True
-    parser.fill_nan_with = "N/A"
-    result = parser.parse()
-
-    # If parsing fails, check the error message
-    if not result:
-        assert parser.error_message is not None
-        print(f"Error message: {parser.error_message}")
-        # Skip the rest of the test
-        return
-
-    assert parser.parsed_dict is not None
-    assert parser.parsed_dict["csv_rows"][1]["age"] == "N/A"
-
-
-@pytest.mark.unit
-def test_fill_nan_with() -> None:
-    """Test the fill_nan_with property and setter."""
-    # Create a CSV file with NaN values
-    content = b"name,age\nAlice,30\nBob,\nCharlie,35"
-    config_file = BytesIO(content)
-    config_file.name = "config.csv"
-
-    # Setting fill_nan_with without enabling fill_nan should have no effect
-    parser = ConfigParser(config_file)
-    parser.fill_nan_with = "Unknown"
-    assert parser.parse() is True
-    assert parser.parsed_dict is not None
-    assert math.isnan(parser.parsed_dict["csv_rows"][1]["age"])
-
-    # Reset and enable both options
-    config_file.seek(0)
-    parser = ConfigParser(config_file)
-    parser.enable_fill_nan = True
-    parser.fill_nan_with = "Unknown"
-    result = parser.parse()
-
-    # If parsing fails, check the error message
-    if not result:
-        assert parser.error_message is not None
-        print(f"Error message: {parser.error_message}")
-        # Skip the rest of the test
-        return
-
-    assert parser.parsed_dict is not None
-    assert parser.parsed_dict["csv_rows"][1]["age"] == "Unknown"
-
-    # Test with numeric fill value
-    config_file.seek(0)
-    parser = ConfigParser(config_file)
-    parser.enable_fill_nan = True
-    parser.fill_nan_with = "0"  # Use string "0" instead of integer 0
-    result = parser.parse()
-
-    # If parsing fails, check the error message
-    if not result:
-        assert parser.error_message is not None
-        print(f"Error message: {parser.error_message}")
-        # Skip the rest of the test
-        return
-
-    assert parser.parsed_dict is not None
-    assert parser.parsed_dict["csv_rows"][1]["age"] == "0"
-
-
-@pytest.mark.unit
-@pytest.mark.parametrize(
-    ("csv_rows_name", "enable_fill_nan", "fill_nan_with", "expected_dict"),
-    [
-        # Default settings
-        ("csv_rows", False, "", {"csv_rows": [{"name": "Alice", "age": 30}, {"name": "Bob", "age": None}, {"name": "Charlie", "age": 35}]}),
-        # Custom row name
-        ("people", False, "", {"people": [{"name": "Alice", "age": 30}, {"name": "Bob", "age": None}, {"name": "Charlie", "age": 35}]}),
-        # Fill NaN with empty string
-        ("csv_rows", True, "", {"csv_rows": [{"name": "Alice", "age": 30}, {"name": "Bob", "age": ""}, {"name": "Charlie", "age": 35}]}),
-        # Fill NaN with custom string
-        (
-            "csv_rows",
-            True,
-            "Unknown",
-            {"csv_rows": [{"name": "Alice", "age": 30}, {"name": "Bob", "age": "Unknown"}, {"name": "Charlie", "age": 35}]},
-        ),
-        # Fill NaN with number as string
-        ("csv_rows", True, "0", {"csv_rows": [{"name": "Alice", "age": 30}, {"name": "Bob", "age": "0"}, {"name": "Charlie", "age": 35}]}),
-        # Combination of custom row name and NaN filling
-        ("people", True, "N/A", {"people": [{"name": "Alice", "age": 30}, {"name": "Bob", "age": "N/A"}, {"name": "Charlie", "age": 35}]}),
-    ],
-)
-def test_csv_options_combined(
-    csv_rows_name: str,
-    enable_fill_nan: bool,
-    fill_nan_with: str,  # Changed from Union[str, int] to just str
-    expected_dict: Dict[str, Any],
-) -> None:
-    """Test combinations of CSV options."""
-    # Create a CSV file with NaN values
-    content = b"name,age\nAlice,30\nBob,\nCharlie,35"
-    config_file = BytesIO(content)
-    config_file.name = "config.csv"
-
-    # Configure parser with the specified options
-    parser = ConfigParser(config_file)
-    parser.csv_rows_name = csv_rows_name
-    parser.enable_fill_nan = enable_fill_nan
-    parser.fill_nan_with = fill_nan_with
-
-    # Parse and verify results
-    result = parser.parse()
-
-    # If parsing fails, check the error message and skip the test
-    if not result:
-        assert parser.error_message is not None
-        print(f"Error message: {parser.error_message}")
-        pytest.skip(f"Parsing failed with error: {parser.error_message}")
-
-    assert parser.parsed_dict is not None
-
-    # For NaN values, we need special handling in the comparison
-    result_dict = parser.parsed_dict
-    expected_rows = expected_dict[csv_rows_name]
-    result_rows = result_dict[csv_rows_name]
-
-    assert len(result_rows) == len(expected_rows)
-
-    for i, (expected_row, result_row) in enumerate(zip(expected_rows, result_rows, strict=False)):
-        for key, expected_value in expected_row.items():
-            if expected_value is None and not enable_fill_nan:
-                # If NaN filling is disabled and expected value is None, check for NaN
-                assert key in result_row
-                assert math.isnan(result_row[key]) or result_row[key] is None
-            else:
-                # Otherwise, compare values directly
-                assert result_row[key] == expected_value, f"Row {i}, key {key}: expected {expected_value}, got {result_row[key]}"
